+++ examples/04-transfer-nft.ts
@@ -2,6 +2,7 @@ import { ethers } from 'ethers';
 import { PrimeSdk } from '../src';
 import { printOp } from '../src/sdk/common/OperationUtils';
 import * as dotenv from 'dotenv';
+import { sleep } from '../src/sdk/common';
 
 dotenv.config();
 
@@ -34,17 +35,22 @@ async function main() {
   console.log('transactions: ', userOpsBatch);
 
   // sign transactions added to the batch
-  const op = await primeSdk.sign();
-  console.log(`Signed UserOp: ${await printOp(op)}`);
+  const op = await primeSdk.estimate();
+  console.log(`Estimated UserOp: ${await printOp(op)}`);
 
-  // sending to the bundler...
+  // sign the userOps and sending to the bundler...
   const uoHash = await primeSdk.send(op);
   console.log(`UserOpHash: ${uoHash}`);
 
   // get transaction hash...
   console.log('Waiting for transaction...');
-  const txHash = await primeSdk.getUserOpReceipt(uoHash);
-  console.log('\x1b[33m%s\x1b[0m', `Transaction hash: ${txHash}`);
+  let userOpsReceipt = null;
+  const timeout = Date.now() + 60000; // 1 minute timeout
+  while((userOpsReceipt == null) && (Date.now() < timeout)) {
+    await sleep(2);
+    userOpsReceipt = await primeSdk.getUserOpReceipt(uoHash);
+  }
+  console.log('\x1b[33m%s\x1b[0m', `Transaction Receipt: `, userOpsReceipt);
 }
 
 main()